requirements.txt
alpaca-trade-api>=3.2
websockets>=12.0
httpx>=0.27
pydantic>=2.8
python-dotenv>=1.0
pandas>=2.1
numpy>=1.26
statsmodels>=0.14
streamlit>=1.36
plotly>=5.24
uvloop; platform_system!="Windows"
